script/requirements.txt
fastapi
uvicorn
python-whois
requests
bs4